% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/links.R
\name{add_all_links}
\alias{add_all_links}
\title{Automatic add function links to RDocumentation to course (beta)}
\usage{
add_all_links(backup_ext = "_back_no_links", autoloads = c("base",
  "methods", "datasets", "utils", "grDevices", "graphics", "stats"),
  course_yml = "course.yml")
}
\arguments{
\item{backup_ext}{The extension to the name of the backup files. Default is "_back_no_links".}

\item{autoloads}{The libraries that are autoloaded into R.
Default is c("base", "methods", "datasets", "utils", "grDevices", "graphics", "stats").}

\item{course_yml}{The name of the course yaml file. Default is "course.yml".}
}
\description{
The \code{add_all_links} function will use the course yaml file to add all links to quoted functions. It will
automatically search for the relevant RDocumentation link.
}
\details{
The function will first look within relevant libraries (the ones manually loaded in the chapter file). Afterwards
it will look in autoloaded libraries and finally it will look in all installed libraries for the relevant
help file link.

IMPORTANT: This function will create backup files, as it overwrites the original chapter file. The automatic link generation
can be undone using the \code{undo_all_links} function. But only use this function right after generating the links,
as it uses the generated backup files. Changes done to the chapter file between using \code{add_all_links} and
\code{undo_all_links} will be undone.
}

