% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/course_file.R
\name{add_chapter_to_course_file}
\alias{add_chapter_to_course_file}
\title{Add a chapter to the list of chapters in the course.yml file}
\usage{
add_chapter_to_course_file(chapter_file_name, chapter_id)
}
\arguments{
\item{chapter_file_name}{the chapter file name to add}

\item{chapter_id}{the id to assign to the chapter inside the course.yml file}
}
\description{
Add a chapter to the list of chapters in the course.yml file
}

