% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/course_yml.R
\name{add_id_to_course_yml}
\alias{add_id_to_course_yml}
\title{Add a course_id to the course.yml file}
\usage{
add_id_to_course_yml(course_id)
}
\arguments{
\item{course_id}{id to be added}
}
\description{
Add a course_id to the course.yml file
}

