\name{author_chapter}
\alias{author_chapter}
\title{Create a new chapter}
\usage{
author_chapter(chapter_name = NULL)
}
\arguments{
  \item{chapter_name}{Character with the name of the
  chapter you'd like to create.}
}
\description{
Creates an R Markdown file for a new course chapter in the
current working directory. The R markdown file already
contains a template which is opened for editing.
}
\examples{
\dontrun{
author_chapter("chapter2")
}
}

