% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/check.R
\name{check_chapters}
\alias{check_chapters}
\title{Check the filenames of the different chapters}
\usage{
check_chapters(course)
}
\arguments{
\item{course}{the course object, containing chapter info, to be checked}
}
\description{
Check the filenames of the different chapters
}

