% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/check_course_object.R
\name{check_course_object}
\alias{check_course_object}
\title{Check the course object}
\usage{
check_course_object(course)
}
\arguments{
\item{course}{The course object to be checked}
}
\description{
Check the course object
}

