% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_leftovers}
\alias{clean_leftovers}
\title{delete the .md and .html files}
\usage{
clean_leftovers(input_file)
}
\arguments{
\item{input_file}{The chapter fie for which the cleaning is needed}
}
\description{
delete the .md and .html files
}

