% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{convert_to_named_list}
\alias{convert_to_named_list}
\title{Convenience function for challenges}
\usage{
convert_to_named_list(instructions)
}
\arguments{
\item{instructions}{unordered list of instructions}
}
\description{
Convenience function for challenges
}

