\name{datacamp_login}
\alias{datacamp_login}
\title{Log in to DataCamp.com via R}
\usage{
datacamp_login()
}
\value{
No return values.
}
\description{
To be able to upload your course to the DataCamp platform,
you need to log in to set-up the connection. The function
will prompt for your DataCamp username and password, and
will then log you into the DataCamp server. Optionally, a
subdomain can be specified (the default is
www.DataCamp.com). Note, in addition to the log in via R,
it is also necessary to log into DataCamp.com via your
browser.
}

