\name{datacamp-package}
\alias{datacamp-package}
\alias{datacamp}
\docType{package}
\title{Tools for the using the interactive www.datacamp.com learning platform}
\description{
Visit www.datacamp.com for more information!
}
\details{
\tabular{ll}{
Package: \tab datacamp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-18\cr
License: \tab What license is it under?\cr
}
}
\author{Jonathan Cornelissen <Jonathan@datacamp.com>}
\references{www.datacamp.com}
\keyword{ package }
