% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparse_chapter.R
\name{deparse_chapter}
\alias{deparse_chapter}
\title{Based on the structured list representing a chapter file, create the chapter file again.}
\usage{
deparse_chapter(chapter_list, chapter_file)
}
\arguments{
\item{chapter_list}{structured list resulting from \link{parse_chapter}}

\item{chapter_file}{Path of the file to which you want to write the list representing the chapter.}
}
\description{
Based on the structured list representing a chapter file, create the chapter file again.
}
\examples{
\dontrun{
chapter_list <- parse_chapter("path_to_file", htmlify = FALSE, check = FALSE)
deparse_chapter(chapter_list, "chapterX.Rmd")
}

}

