% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{doc_url}
\alias{doc_url}
\title{Documentation path}
\format{\preformatted{ chr "http://docs.datacamp.com/teach"
}}
\usage{
doc_url
}
\description{
Documentation path
}
\keyword{datasets}

