% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_code}
\alias{extract_code}
\title{Extract clean code from html}
\usage{
extract_code(html)
}
\arguments{
\item{html}{The html content to 'codify'}
}
\description{
Extract clean code from html
}

