% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_html}
\alias{extract_html}
\title{Extract html from a chunk of text}
\usage{
extract_html(x)
}
\arguments{
\item{x}{Raw text to be converted to html}
}
\description{
If a second-level header is in the content, this is removed.
}

