% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_markdown}
\alias{extract_markdown}
\title{Extract R Markdown}
\usage{
extract_markdown(x, default_name)
}
\arguments{
\item{x}{The html content to 'rmarkdown codify'}

\item{default_name}{the default file name of an r markdown code chunk, if no title is specified in three curly braces.}
}
\description{
Extract R Markdown
}

