% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_mc}
\alias{extract_mc}
\title{Function to create a vector with multiple choice options}
\usage{
extract_mc(x)
}
\arguments{
\item{x}{raw text that should contain an unordered list.}
}
\description{
Function to create a vector with multiple choice options
}

