% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_named_list}
\alias{extract_named_list}
\title{Convenience function for challenges}
\usage{
extract_named_list(x)
}
\arguments{
\item{x}{input}
}
\description{
Returns embedded list. Titles are second-level headers, the content beneath it the content.
}

