% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_skills}
\alias{extract_skills}
\title{Extract skills from an exercise}
\usage{
extract_skills(x)
}
\arguments{
\item{x}{content of the skills part (should be comma separated skill ids)}
}
\description{
Extract skills from an exercise
}

