% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_title}
\alias{extract_title}
\title{Extract the title}
\usage{
extract_title(x)
}
\arguments{
\item{x}{The main content containing a second-level header title}
}
\description{
Extract the title
}

