% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_video_link}
\alias{extract_video_link}
\title{Extract video link}
\usage{
extract_video_link(x)
}
\arguments{
\item{x}{content of the video link part}

\item{lang}{lang part}
}
\description{
Extract video link
}

