% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{fix_specials}
\alias{fix_specials}
\title{Helper function to convert the code words _tbt_ and _tast_ to triple backticks and triple asterisks respectively
This is to escape from the backtick inception in R markdown files (R markdown containing R markdown...)}
\usage{
fix_specials(code)
}
\arguments{
\item{code}{The code to convert}
}
\description{
Helper function to convert the code words _tbt_ and _tast_ to triple backticks and triple asterisks respectively
This is to escape from the backtick inception in R markdown files (R markdown containing R markdown...)
}

