% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/course_yml.R
\name{get_chapter_id}
\alias{get_chapter_id}
\title{Get the chapter id from the course.yml for a particular chapter file.}
\usage{
get_chapter_id(file_name)
}
\arguments{
\item{file_name}{the chapter file name to get the id for}
}
\description{
Get the chapter id from the course.yml for a particular chapter file.
}

