% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/course_yml.R
\name{get_chapter_index}
\alias{get_chapter_index}
\title{Get the chapter id from the course.yml for a particular chapter file.}
\usage{
get_chapter_index(chapter_file_name)
}
\arguments{
\item{chapter_file_name}{the chapter file name to get the id for}

\item{course}{parsed course information from course.yml}
}
\description{
Get the chapter id from the course.yml for a particular chapter file.
}

