% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{have_a_look}
\alias{have_a_look}
\title{Information message}
\format{\preformatted{ chr "Have a look at the documentation on http://teach.datacamp.com."
}}
\usage{
have_a_look
}
\description{
Information message
}
\keyword{datasets}

