% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{html2txt}
\alias{html2txt}
\title{Convenience function to convert html codes:}
\usage{
html2txt(str)
}
\arguments{
\item{str}{html string to convert back to more human readable characters}
}
\description{
Convenience function to convert html codes:
}

