% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/course_yml.R
\name{load_course_file}
\alias{load_course_file}
\title{Load course file and check if all necessary info is there and is valid}
\usage{
load_course_file()
}
\description{
Load course file and check if all necessary info is there and is valid
}

