% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/course_yml.R
\name{load_course_yml}
\alias{load_course_yml}
\title{Load course.yml file and check if all necessary info is there and is valid}
\usage{
load_course_yml()
}
\description{
Load course.yml file and check if all necessary info is there and is valid
}

