% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_multiple_choice_vector}
\alias{make_multiple_choice_vector}
\title{Function to create a vector with the multiple choice options:}
\usage{
make_multiple_choice_vector(instructions)
}
\arguments{
\item{instructions}{html code of an unordered list}
}
\description{
Function to create a vector with the multiple choice options:
}

