% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/render_chapter_json.R
\name{render_chapter_json_for_datacamp}
\alias{render_chapter_json_for_datacamp}
\title{Create a JSON from the slidified chapter}
\usage{
render_chapter_json_for_datacamp(file_name, payload, force, skip_validation)
}
\arguments{
\item{file_name}{name of the chapter file}

\item{payload}{structure built by slidify package, containing all data in the chapter r markdown file.}

\item{force}{whether or not to impose the local structure onto the remote datacamp course.}

\item{skip_validation}{whether or not to skip validity of uploaded exercises on DataCamp's servers.}
}
\description{
Create a JSON from the slidified chapter
}

