% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/build_capstone_network.R
\name{show_capstone_graph}
\alias{show_capstone_graph}
\title{Build graph structure of Capstone chapter}
\usage{
show_capstone_graph(chapter_file)
}
\arguments{
\item{chapter_file}{path to the chapter file you want to parse}
}
\description{
Build graph structure of Capstone chapter
}

