% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utils.R
\name{split_lines}
\alias{split_lines}
\title{split string into separate lines (copied from knitr)}
\usage{
split_lines(x)
}
\arguments{
\item{x}{string or file to split in lines}
}
\description{
split string into separate lines (copied from knitr)
}

