% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links.R
\name{undo_all_links}
\alias{undo_all_links}
\title{Undo all link creation using the backup files you created (beta)}
\usage{
undo_all_links(backup_ext = "_back_no_links", course_yml = "course.yml")
}
\arguments{
\item{backup_ext}{The extension to the name of the backup files. Default is "_back_no_links".}

\item{course_yml}{The name of the course yaml file. Default is "course.yml".}
}
\description{
The \code{undo_all_links} function should only be used right after using the \code{add_all_links} function.
If you inspect the code and the link generation seem to have messed up the course format.
This should not happen, but everything can happen in a beta version. Needs more testing.
}

