\name{upload_all_chapters}
\alias{upload_all_chapters}
\title{Upload all chapters}
\usage{
upload_all_chapters()
}
\value{
No return values.
}
\description{
Loop over all \code{.Rmd} files in the course directory and
upload them using the \code{upload_chapter()} function.
}
\examples{
\dontrun{
upload_all_chapters()
}
}

