% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/upload_chapter.R
\name{upload_chapter_json}
\alias{upload_chapter_json}
\title{Upload the chapter json}
\usage{
upload_chapter_json(chapter_json, chapter_file, open = TRUE)
}
\arguments{
\item{chapter_json}{the JSON string to be posted}

\item{chapter_file}{chapter file name that is being uploaded}

\item{open}{whether or not to open the teach website after upload.}
}
\description{
Upload the chapter json
}

