% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/upload_course.R
\name{upload_course_json}
\alias{upload_course_json}
\title{Upload the course json}
\usage{
upload_course_json(theJSON, open = TRUE)
}
\arguments{
\item{theJSON}{the JSON string to be posted}

\item{open}{whether or not to open the teach website after upload.}
}
\description{
Upload the course json
}

