% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/author.R
\name{add_exercise}
\alias{add_exercise}
\title{Add exercise to chapter file}
\usage{
add_exercise(chapter_file, language = c("r", "python"),
  type = c("NormalExercise", "MultipleChoiceExercise", "VideoExercise"),
  simplified = FALSE)
}
\arguments{
\item{chapter_file}{path to the chapter you want to append a template to}

\item{type}{type of the exercise you want to add a template for}

\item{simplified}{whether or not to add a simplified template (only TRUE is supported for the moment)}
}
\description{
Add exercise to chapter file
}

