% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/author.R
\name{add_exercise}
\alias{add_exercise}
\title{Add exercise to chapter file}
\usage{
add_exercise(chapter_file, lang, type = c("NormalExercise",
  "MultipleChoiceExercise", "VideoExercise"), simplified, title = NULL,
  content = NULL)
}
\arguments{
\item{chapter_file}{path to the chapter you want to append a template to}

\item{lang}{The programming language of the exercises in the template}

\item{type}{type of the exercise you want to add a template for (default is "NormalExercise")}

\item{simplified}{Whether or not to build simplified templates.}

\item{title}{Title of the exercise}

\item{content}{Short summary of the exercise (will be filled in in assignment part)}
}
\description{
Add exercise to chapter file
}

