% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author.R
\name{add_exercise}
\alias{add_exercise}
\title{Add exercise to chapter file}
\usage{
add_exercise(chapter_file, lang, type = c("NormalExercise",
  "MultipleChoiceExercise", "VideoExercise", "RStudioMultipleChoiceExercise",
  "ShinyNormalExercise"), title = NULL, content = NULL)
}
\arguments{
\item{chapter_file}{path to the chapter you want to append a template to}

\item{lang}{The programming language of the exercises in the template}

\item{type}{type of the exercise you want to add a template for (default is "NormalExercise")}

\item{title}{Title of the exercise}

\item{content}{Short summary of the exercise (will be filled in in assignment part)}
}
\description{
Add exercise to chapter file
}
\examples{
\dontrun{
add_exercise("chapter1.Rmd", lang = "r", type = "NormalExercise")
add_exercise("chapter1.Rmd", lang = "r", type = "MultipleChoiceExercise")
add_exercise("chapter1.Rmd", lang = "r", type = "VideoExercise")
add_exercise("chapter1.Rmd", lang = "r", type = "NormalExercise",
             title = "My Normal Exercise", content = "This is a normal exercise.")
add_exercise("chapter1.Rmd", lang = "r", type = "MultipleChoiceExercise")
             title = "My Multiple Choice Exercise", content = "This is an MCE.")
add_exercise("chapter1.Rmd", lang = "r", type = "VideoExercise",
             title = "My Video", content = "This video discusses interesting things.")

add_exercise("chapter1.md", lang = "python", type = "NormalExercise")
add_exercise("chapter1.md", lang = "python", type = "MultipleChoiceExercise")
add_exercise("chapter1.md", lang = "python", type = "VideoExercise")
add_exercise("chapter1.md", lang = "python", type = "NormalExercise",
             title = "My Normal Exercise", content = "This is a normal exercise.")
add_exercise("chapter1.md", lang = "python", type = "MultipleChoiceExercise")
             title = "My Multiple Choice Exercise", content = "This is an MCE.")
add_exercise("chapter1.md", lang = "python", type = "VideoExercise",
             title = "My Video", content = "This video discusses interesting things.")
}

}

