% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/author.R
\name{author_chapter}
\alias{author_chapter}
\title{Generate course and chapter template}
\usage{
author_chapter(lang = c("r", "python"), simplified = FALSE, open = TRUE)
}
\arguments{
\item{lang}{The programming language of the exercises in the template exercises (default is "r")}

\item{simplified}{Whether or not to build a simplified course scaffold (default is FALSE)}

\item{open}{Whether or not to open the files in a text editor after creating it.}
}
\description{
The \code{author_course} function will:
\enumerate{
 \item create and open a `course.yml` file, with course information
 \item create and open a `chapter1.md` file, containing some exercise templates
}
}
\examples{
\dontrun{
author_chapter(lang = "python", simplified = TRUE)
author_chapter(lang = "python", simplified = FALSE)
author_chapter(lang = "r", simplified = TRUE)
author_chapter(lang = "r", simplified = FALSE)
}
}

