% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author.R
\name{author_chapter}
\alias{author_chapter}
\title{Generate chapter file}
\usage{
author_chapter(lang, title = NULL, description = NULL, internal = FALSE)
}
\arguments{
\item{lang}{Programming language of the chapter}

\item{title}{Title of the the chapter (optional)}

\item{description}{Description of the chapter (optional)}

\item{internal}{Do not touch this parameter (for internal use)}
}
\description{
Generate a chapter file with 3 exercise templates so you can start right away.
}
\examples{
\dontrun{
author_chapter(lang = "python")
author_chapter(lang = "r")
author_chapter(lang = "r",
               title = "My Chapter",
               description = "This is my chapter!")
}

}

