% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/author.R
\name{author_chapter}
\alias{author_chapter}
\title{Generate chapter file}
\usage{
author_chapter(lang, simplified = FALSE, title = NULL, description = NULL,
  internal = FALSE)
}
\arguments{
\item{lang}{Programming language of the chapter}

\item{simplified}{Whether or not to build simplified templates.}

\item{title}{Title of the the chapter (optional)}

\item{description}{Description of the chapter (optional)}

\item{internal}{Do not touch this parameter (for internal use)}
}
\description{
Generate a chapter file with 3 exercise templates so you can start right away.
}
\examples{
\dontrun{
author_chapter(lang = "python", simplified = TRUE)
author_chapter(lang = "python", simplified = FALSE)
author_chapter(lang = "r", simplified = TRUE)
author_chapter(lang = "r", simplified = FALSE)
}
}

