% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/author.R
\name{author_course}
\alias{author_course}
\title{Generate course and chapter scaffold}
\usage{
author_course(language = c("r", "python"), simplified = FALSE,
  open = TRUE)
}
\arguments{
\item{language}{The programming language of the course you want to build.}

\item{simplified}{Whether or not to build a simplified course scaffold.}

\item{open}{Whether or not to open the file in a text editor after creating it.}
}
\value{
No return values.
}
\description{
The \code{author_course} function will:
\enumerate{
 \item create and open a `course.yml` file, a scaffold to provide the necessary course information.
 \item create and open a `chapter1.md` file, a scaffold for creating your first chapter.
}
}
\details{
In the `course.yml` file a course title, course author and course description should be provided. Next, it also contains the unique
course ID, and a list of the chapters within the course. The `chapter1.md` file provides the structure and building blocks of the
first chapter.
}
\examples{
\dontrun{
author_course()
}
}

