% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/author.R
\name{author_course}
\alias{author_course}
\title{Generate course and chapter template}
\usage{
author_course(lang)
}
\arguments{
\item{lang}{The programming language of the exercises in the template}
}
\description{
The \code{author_course} function will:
\enumerate{
 \item create and open a `course.yml` file, with course information
 \item create and open a `chapter1.md` file, containing some exercise templates
}
}
\examples{
\dontrun{
author_course(lang = "python")
author_course(lang = "r")
}
}

