\name{datacamp_login}
\alias{datacamp_login}
\title{Log in to datacamp.com}
\usage{
datacamp_login()
}
\value{
No return values.
}
\description{
Used to log in to datacamp.com. The function will prompt
for your username and password and log you in to the
datacamp server. Optionally a subdomain can be specified
(the default is www.datacamp.com). Note that you must also
log into datacamp.com with your browser.
}

