% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_code}
\alias{extract_code}
\title{Extract R code chunks from raw text}
\usage{
extract_code(x)
}
\arguments{
\item{x}{Raw text containing R code chunk}
}
\description{
Extract R code chunks from raw text
}

