% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_code}
\alias{extract_code}
\title{Extract code chunks from raw text}
\usage{
extract_code(x, lang = "r")
}
\arguments{
\item{x}{Raw text containing R code chunk}

\item{lang}{The programming language}
}
\description{
Extract code chunks from raw text
}

