% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helpers.R
\name{extract_markdown}
\alias{extract_markdown}
\title{Extract the R Markdown code files}
\usage{
extract_markdown(html, default_name)
}
\arguments{
\item{html}{The html content to 'rmarkdown codify'}

\item{default_name}{the default file name of an r markdown code chunk, if no title is specified in three curly braces.}
}
\description{
Extract the R Markdown code files
}

