% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/course_file.R
\name{get_chapter_index}
\alias{get_chapter_index}
\title{Get the chapter id from the course.yml for a particular chapter file.}
\usage{
get_chapter_index(chapter_file_name)
}
\arguments{
\item{chapter_file_name}{the chapter file name to get the id for}
}
\description{
Get the chapter id from the course.yml for a particular chapter file.
}

