% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/chapter_parser.R
\name{parse_chapter}
\alias{parse_chapter}
\title{Parse chapter file into R list}
\usage{
parse_chapter(chapter_file)
}
\arguments{
\item{path}{to the chapter file you want to parse}
}
\description{
Parse chapter file into R list
}

