% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/upload_chapter.R
\name{upload_all_chapters}
\alias{upload_all_chapters}
\title{Upload all chapters in the current directory}
\usage{
upload_all_chapters(force = FALSE, open = TRUE)
}
\arguments{
\item{force}{whether or not to force (look at documentation upload_chapter)}
}
\description{
Chapter files should be names chapterX.Rmd. Buggy convenience function.
}

