% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/upload_chapter.R
\name{upload_chapter}
\alias{upload_chapter}
\title{Create or update a chapter}
\usage{
upload_chapter(chapter_file = "chapter1.Rmd")
}
\arguments{
\item{chapter_file}{path to the ".Rmd" file to be uploaded}

\item{force}{boolean, FALSE by default, specifies whether exercises should be removed. If set, will prompt for confirmation.}

\item{open}{boolean, TRUE by default, determines whether a browser window should open, showing the course creation web interface}

\item{ask}{boolean, TRUE by default, determines whether you are asked for confirmation if you set force to TRUE.}
}
\value{
No return values.
}
\description{
Create or update a chapter
}
\examples{
\dontrun{
# Upload without possibly deleting existing exercises
upload_chapter("chapter1.Rmd")

# Completely sync online chapter with your local markdown file
upload_chapter("chapter1.Rmd", force = TRUE)
}
}

