% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/upload_course.R
\name{upload_course}
\alias{upload_course}
\title{Create or update a course}
\usage{
upload_course(course_filder = ".", open = TRUE, force = FALSE)
}
\arguments{
\item{course_folder}{character, denoting where the course.yml file is located}

\item{open}{boolean, TRUE by default, determines whether a browser window should open, showing the course creation web interface}

\item{force}{boolean, FALSE by default, that allows to remove chapters from the live course that are not in the course.yml}
}
\description{
Uploads the \code{course.yml} file to datacamp.com. Use this function to change the course title, description, etc. and to update the chapters' ordering.
}
\details{
If you're not yet logged in when calling this function, you'll be prompted to log in.
}
\examples{
\dontrun{
upload_course()
}
}

